requirements.txt
anthropic==0.49.0
pdfplumber==0.11.0
pypdf==4.3.1
sqlite-utils==3.36
python-dotenv==1.0.1
